+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.4.13 - 23 Oct 2025
+- [#1502](https://github.com/elysiajs/elysia/issues/1502) afterHandle doesn't update status
 - [#1495](https://github.com/elysiajs/elysia/pull/1495) request server hook parameters are typed as any (Bun 1.3.0)
 - [#1483](https://github.com/elysiajs/elysia/pull/1483) handle standard schema validators in ValidationError.all
 - [#1459](https://github.com/elysiajs/elysia/pull/1459) fix strictPath behavior when aot: false is set